Move end-of-rib timeout value from schema to contrail-control.conf
Similar to ifmap-end-of-rib timeout, BGP and XMPP end-of-rib-timeout values
shall be configured/tuned via /etc/contrail/contrail-control.conf. They shall
default to 30 seconds

When the timer is running, timer is reposted if new subscribe/unsubsribe
requests are received. This ensures that we do not prematurely exit out of
GR, when startup activities are still in progress.

EndOfRib receive event inferring shall be further tuned in subsequent commits

Change-Id: I0cda0dfcd58facba2eed01330b19308a9bec1a0f
Partial-Bug: #1537933