Treat VN index >= 8000000 as global
This is required for correct operation in multi-region scenarios with
service chaining. If there are 2 or more regions re-originating routes
for service chaining, we do not want to derive the origin vn from the
route targets of routes received from other regions. We want the global
controller to manage allocation of vn ids and trust received values if
they have been allocated by global controller.

Values greater than 8000000 are treated as global and are allocated by
the global controller. Need to disregard the ASN and compare only the
vn ids in this case.

Change-Id: I2404517b5d0024ebbea04eb053273aea45ddf20d
Closes-Bug: 1612415