* Pick covering route irrespective of whether nexthop is resolved or not
Assume there are three route
1> 1.1.1.0/32 with nexthop of intf1
2> 1.1.1.0/24 with nexthop of intf1
3> 0.0.0.0/0 with nexthop of MX

If intf1 is deleted, then its corresponding nexthop would also be
deleted, hence in ksync if route first route notified for delete
is 1.1.1.0/32 replacement route for it would be set as 0/0 even
though 1.1.1.0/24 is not deleted in kernel. Removing the check
to verify if nexthop is valid since we only need the route resolved
state.
Partial-bug:#1605748

Change-Id: I5d3c02d2521b7c0c113c6489cd82e2e141560911