* Compare only Tunnel SIP or VMI while fetching component index.
If AAP is specified with vMAC, new label would allocated to
the route, to fetch component NH for composite NH dummy component
NH is allocated which matches for both label and NH, this strict
check for label is not needed. Removing the check and replacing
with check for VMI or tunnel IP.
Test case for same.
Closes-bug:#1615571

Change-Id: I1e48a03a585a8892e0fa71d03b7c4e3df8a97368