Infra to disable TaskGroup/TaskEntry for UT
Task library to be enhanced for disabling a specific
TaskGroup/TaskEntry for UT purposes. New Task will not
be run immediately by TaskScheduler if it matches TaskGroup/TaskEntry
which is disabled and will be put in waitq. Enqueued tasks will be
run by scheduler after respective TaskGroup/TaskEntry is enabled again.

Change-Id: I1d2114bb2a42cd0261404cd17722b568ee22bdd8
Closes-bug: #1463293