Improve ifmap graph walk performance
Following areas in boost graph implementation for ifmap db needs optimization
1. BFS: Current BFS visits all vertices in DBGraph before starting the BFS.

2. BFS applies EdgeFilter and NodeFilter on all edges and nodes encountered in
the BFS walk. E.g. if there are 1K VMI in a VN, then during the BFS walk of the
VM, VN node is encountered due to VN<->VMI<->VM link. When walking edges of the
VN, all the VMIs belonging to the VN applied with EdgeFilter and NodeFilter.
Since it is not part of whitelist, other VMIs are not visited.

3. During the processing of Vertex during graph walk, JoinVertex, all edges are
notified. This walk of all edges from vertex is expensive in case VN node in the
above example.(1K VMI from a VN). While visiting the VN, all VMI links are
unnecessarily notified.

To solve the above issues, following optimizations are implemented in this commit

1. Store the out edges of the Vertex in multiset sorted by metadata name[i.e.
name of the link]. This helps in iterating the edges of the vertex better.

2. Implement BFS in CN to avoid iterating all vertices in graph to color it.
The color is implemented as genid stored in each vertex.
The BFS implementation also doesn't visit all edges of the graph and apply the
filter. Since the out edges of the vertex is stored in multiset, we can visit
edges only present in the whitelist.  With this, BFS can skip over uninteresting
links in logN time instead of N.

3. The whitelist is stored such that for a given vertex, set of all interested
edges (metadata name is stored). Please note, allowed edge is sorted by edge name
and out edge list is also sorted by edge name. So out edge list walk can be optimized
by using lower_bound()

4. Edges from a visited Vertex is not notified in the JoinVertex(). Instead only
interested edges are notified by calling Notify() when an Edge is visited in graph
walk.

Change-Id: Iadd815d6897dc84c786c7079659023e190e5ef20
Closes-Bug: #1619943