Reduce memory allocation overhead in update generation code
Use static vectors of string and xml_document, 1 per BgpSenderPartition,
instead of creating and destroying them dynamically as BgpXmppMessages
are created and destroyed.

The vectors have an entry for each shard. This reduces memory allocation
overhead by allowing BgpXmppMessage allocated from a BgpSenderPartition
to use the string and xml_document associated with it's shard instead of
creating and destroying them repeatedly.

Note that pugixml library allocates memory for a document in increments
of 32KB pages and then manages smaller allocations (nodes or attributes)
using these pages.  Pre-creating the xml_documents ensures that pugixml
does only a single 32KB allocation per document and then recycles the
same memory when building the tree for each route/item.  Without this
optimization, we would allocate/free a 32KB page for each route/item.

Similarly, using a static vector of strings allows the same string to be
re-used by the BgpSenderPartition for a given shard instead of growing
the capacity of the string as routes are added to a BgpXmppMessage and
then destroying the string along with the BgpXmppMessage.

Change-Id: I8fb5d79c3d015c330bbc6409b8cd6180b2025479
Partial-Bug: 1607132