Send admin state and all the IPs of VMI in VMI UVE
Admin state is required to raise alarm when VMI is down but not admin-down.
Also send all the IPs that a VMI has instead of sending only primary IP.
Removed unused variable flow_tree_ in FlowMgmtManager
Also change FlowEntryPtr in FlowExportInfo to FlowEntry* as reference is already held by key.

Change-Id: I10457559ebe4cca8563ef319818aea1153e6a61d
Closes-Bug: #1620905