Add GR flags to XmppOpenMessage
When an agent (XmppClient) restarts, in current code and in many scenarios such
as reboot, forwarding state may not be preserved. In such cases, when new
XmppConnection is detected, GR (helper) mode must be aborted and routes should
be cleaned up right away

Added a simple "graceful-restart" attribute to Xmpp Open message. If this
attribute is set to "true" only then shall the routes be preserved, otherwise
session shall be cleaned up non-gracefully.

Reorganize code inside XmppStateMachine::ProcessStreamHeaderMessage() to make
it easier to read and understand. If agent restarts without sending attribute
"graceful-restart" or if it sends it with "false" as the value, then
existing connection must be closed non-gracefully, there by aborting the GR.

Ensured that graceful_restart_test did not pass if GR flag is not set by the
mock agents. Once set, the integration test also passed

More tests can be added after this way to advertise GR state is reviewed and
accepted.

Change-Id: I9a7d36e65b950dd57bd8aaedde88ea10cf7fead3
Partial-Bug: #1537933