Add GR flags to XmppOpenMessage
When an agent (XmppClient) restarts, in current code and in many scenarios such
as reboot, forwarding state may not be preserved. In such cases, when new
XmppConnection is detected, GR (helper) mode must be aborted and routes should
be cleaned up right away

Added a simple "gr" attribute to Xmpp Open message. If this attribute is set to
"true" only then shall the routes be preserved, otherwise session shall be
cleaned up non-gracefully.

Ensured that graceful_restart_test did not pass if GR flag is not set by the
mock agents. Once set, the integration test also passed

More tests can be added after this way to advertise GR state is reviewed and
accepted.

Change-Id: I3d4c7debee4e4ff2f49d44a5ab8adb0875f567dc
Partial-Bug: #1537933