Collector stops sending Discovery Publish Message to Discovery Server.
Discovery Client is waiting infinetly for the "http client" task to be scheduled which
periodically sends publish message to Discovery Server.

Both VncApi object and Discovery Client object uses the same "http client" task and instance.
Deletion of either of the object does not cleanup with WorkQueue and hence the "http client"
task scheduled to execute the work on the deleted object does not succeed and waits
forever. Task Scheduler has a task running with the same task-id and instance and  hence
similar tasks are not scheduled.

Shutdown the work_qeueue on HttpClient object delete.

Change-Id: I2787a2594631bc73502bdf7946cf161f90006805
Closes-Bug:1617112