/!\WIP/!\ [VNC API Server] Remove irond dependency
That patch remove dependency to the irond IF-MAP server. It implements a
minimalist IF-MAP v2 server run as a part of the contrail API server.
That server is directly populated by the VNC API server (directly access
the python dict that represents the ifmap graph no HTML serialization).
It keeps a HTML interface for control nodes to poll the graph.

Two servers are started to serve part of the IF-MAP protocol with and
without ssl. Only request 'newSession', 'subscribe' (root graph only) and
'poll' methods are implemented.

=== Results with a db populated with 79440 resources ===
Time elapsed for the vnc api server to sync ifmap db from:
- without patch
    #1: 0:04:00.0272121
    #2: 0:03:51.0994972
    #3: 0:03:49.4030671
- with patch:
    #1: 0:00:39.394018
    #2: 0:00:38.890950
    #3: 0:00:39.109681

Time elapsed for a python script to create a new ifmap session,
subscribe to the root graph and poll it (searchResult):
- without patch
    #1: 0:00:14.655667
    #2: 0:00:13.452200
    #3: 0:00:13.905878
- with patch:
    #1: 0:00:04.591735
    #2: 0:00:04.592040
    #3: 0:00:04.192904

Still work in in progress:
- implement the 'search' method for tools like ifmap-view
- rework unit tests to use that new ifmap implementation

Closes-Bug: #1629268
Change-Id: Ib35b48b20c8d46005bf18e8f9b81064985099ff7