Handle more than 1 level of resolution in PathResolver
The code didn't previously handle the scenario where a BgpRoute with
one or more resolved paths is itself used to resolve ResolverPaths.
This scenario caused a concurrency issue wherein one partition was
attempting to modify a BgpRoute (by adding/deleting resolved paths)
while another partition was trying to access the same BgpRoute (to
get the BgpPaths to use as nexthops).

Fix by introducing a read-write mutex in the ResolverRouteState to
serialize access to the corresponding BgpRoute.  Take a write lock
on the ResolverRouteState for the BgpRoute being modified and a read
lock on the ResolverRouteState for the BgpRoute to the nexthop.

Change-Id: I485e197ee92d4ba2347d136fecaec29b0c068b65
Closes-Bug: 1622618