Make BgpSenderPartition::SetQueueActive more efficient
Have RibOutUpdates::TailDequeue fill in the bitset of unsync peers
based on the tail marker instead  of building it by going through
all peers for the RibOut in BgpUpdateSender::SetQueueActive.

The new approach is more efficient because we only really need to
update the unsync peers based on the tail marker. Going through all
the unsync peers in the RibOut is unnecessary since the rest of the
peers will already have their queue active state updated previously
i.e. when they got blocked.

Consider a scenario where a RibOut has member peers 0-7 and the tail
marker for the RibOut has peers 0-5.  Peers 6-7 got blocked earlier
and have their own marker(s). Further, assume that peers 4-5 are out
of sync i.e. they got blocked while doing TailDequeue for some other
RibOut.

If a TailDequeue is performed for the RibOut, the old code would go
through all the peers to build the sync and unsync bitsets, perform
TailDequeue, and then mark the RibOut as active for all unsync peers
4-7. In the new scheme, TailDequeue would return a bitset with peers
4-5 as unsync peers and BgpSenderPartition::SetQueueActive will only
mark the RibOut as active for peers 4-5. State for peers 6-7 does not
need to be updated since that would have been done at the time they
got blocked when doing TailDequeue for the same RibOut.

Also fix UTs in bgp_update_sender_test appropriately.

Change-Id: I9b567832e87a553af1bdd178bf29b13263526834
Partial-Bug: 1607132