Make BgpSenderPartition::UpdatePeerQueue more efficient
The old implementation used to go through all peers for a RibOut
and build a bitset of send ready peers to pass to PeerDequeue.
The bitset is used to decide which peers to merge with the start
marker when we encounter other markers while doing peer dequeue.

The new implementation determines which peers are send ready more
dynamically i.e. as we encounter other markers during peer dequeue.
This avoids unnecessary work in cases where we do not encounter
many markers.  This could be because some markers are behind the
start marker for the peer or because the all peers in the start
marker get blocked or because some peers are in the tail marker.
Further, the new implementation determines the send ready state
directly from the RibOut instead of the BgpUpdateSender since the
RibOut state is more authoritative.

Also treat send_ready state in PeerState as advisory and send_ready
in the IPeerUpdate as authoritative. This if justified because the
IPeerUpdate could become blocked while processing updates from some
other BgpSenderPartition.

Also fix a few update related UTs appropriately.

Change-Id: Ib40d0a0d71c48f56e1af30962cc6f2459be4140a
Partial-Bug: 1607132