Fetch all query rows before processing
contrail-logs takes lot of time processing the query result. Profiling
of the contrail-logs script showed that lot of time is being spent on
xmltodict.parse() and this impacts the overall time taken to fetch all
the rows from the anlaytics-api. TTL for the query result is set to 5 minutes
in redis. Therefore, if the number of records is more for the specified time
duration in the query, then it is possible that some rows may expire
before they are read and hence contrail-logs would display only partial
query result.
To avert this issue, contrail-logs should fetch the query result
completely before processing it.

Change-Id: Ic54731836675009895d69a26514d5699da96bf73
Closes-Bug: #1632446
(cherry picked from commit 71b508dab0ad2855dfb8327c67812b8e4a604e88)