Update CompositeNH on VMI policy status change.
When label of VMI changes and if that VMI (ie VMI's InterfaceNH) is part of
ECMP, then update the CompositeNH for ECMP route to point to right label for
that VMI. Label of VMI can change when policy-status of VMI changes

Closes-Bug: #1602944

Update Path of static-route with correct label on VMI’s label change.
When VMI’s policy status is changed, its label is modified. The updated
label was not set in static-route’s (associated with VMI) path.

When policy changes, the label of VMI is updated. The updated VMI label was not
reflected in AAP routes. Added UT to verify this.

Remove the API InetUnicastAgentRouteTable::ReEvaluatePaths as we don’t need
separate handling for policy change because we change label on policy change.

Change-Id: If05ea42f150f1c147ec04d6a5da6cc1a1f551113
Closes-Bug: #1603958