Added Port Security Enabled extension
This is a neutron extension that we claim to support, but it is
not currently implemented. To impement this, we will need to
support this flag at network and port level. At the port level,
if port security is disabled, no security groups can be applied
to the port and the port can communicate with all other ports.

At the time of port creation, if a specific value is not provided,
we will copy if from the corresponding network. The default value
for both will be True, i.e., port security is enabled unless
explicitly disabled by the user.

1. Added check 'port_security_enabled' flag at the time of creation and update.
2. If not disabled, then only security group is added
3. At the time of update, it can be disabled if no security group is associated
   or a new security group can be added if 'port_security_enabled' was not disabled
   at the time of port creation.

Change-Id: I7643373210f1a20ee1c6a352a3b3c39c5d3a76c7
Closes-Bug: 1632961