Provide show route filtering support based on source, family and protocol
To current set of show route filters, add source, family and protocol as
additional filters

Override IPeer::ToString() with IPeer::ToUVEKey() in show_route_test in order
to make the IPeer source string unique. In tests, ToString() is not unique when
multiple peering sessions are formed between the same pair of BgpServers.

Change-Id: Ie1589d3c1152bb61278e0a5e347559488ea42993
Partial-Bug: #1595983
Closes-Bug: #1425610