Fix Invalid flow-rate exported from Vrouter UVE
Currently flow-rate is updated in Agent::FlowEvent task while it is being read in Agent::Uve task and there is no exclusion between these tasks.
Made changes to update flow-rate in a new task which has exclusion with Agent::Uve. Also ensured in Agent::Uve task that flow-rate
is not exported if it is not yet computed.

Change-Id: I19bff570b0990a3b932576bf23c4b13a8260a3a6
Closes-Bug: #1639698