Moving TbbKeepAwake task to TaskScheduler
In the current infra, tbb keepawake task has to be instantiated by
each client(agent, control etc) which requires it. It is suggested to
move this part to TaskScheduler so that task is created when it is
instantiated, given EventManager object to it. This timer task is created
with default timeout of 1 sec. Client can modify this value according to its
need.

Change-Id: Idc7b006b8a659a2bf597a37cd91030ccd77d2bfd
Closes-Bug: #1637293