Fixup interface-config module
The CfgIntTable was modelled as DBTable and acted as an intermediate table
to store all port-subscriptions. The processing to add a port is -

1. On receiving PortIpc message, the PortIpc module enqueues request to
add CfgIntEntry
2. The InterfaceCfgClient gets notification of CfgIntEntry and then
enqueues request to InterfaceConfig table to add/delete Oper entry.

Storing port-subscriptions in form of DBTable is not needed. So, replacing
CfgIntTable and InterfaceCfgClient class with PortSubscribeTable.

The Port subscriptions are added to the PortSubscribeTable and in
parallel a Interface table request is enqueued to add/delete
VmInterface.

The PortSubscribeTable also supports multiple type of subscriptions.
Nova style subscriptions pass VMI-UUID and other attributes
In next commits, K8S style subscriptions will have VM-UUID and Network
Labels

Change-Id: I3fef2d878da0c65dd147987bd4c57009ac9799a9
Partial-Bug: #1635923