Use subnet and subnet_list from original address
While creating ACL for service chains, it is possible that the remote
network is specified as wild card. In that case, we need to expand it
to actual VNs while creating ACLs because each will have a different
VRF to be assigned. However, we were not copying CIDRs configured on
the policy rule. With this change, we copy the CIDRs from the policy
rule to ACL rule.

Unit tests will be added for mainline commit.

Change-Id: I6d3f84251f0b96d981f87ce1da79fd1ae9f5f9bb
Closes-Bug: 1641701
(cherry picked from commit 16967e44152bd2b00259e5fb98273df2d7aec5c8)