Fix issue of FlowEntry not getting aged out.
Issue:
FlowStats module has two lists for holding flow entries. One list is sorted on
Flow Pointer (flow_tree_). The other list holds (flow_export_info_list_) flow
entries in the order in which they were sent to FlowStats Module. The
flow_export_info_list_ is used for iterating and aging out flows. In problem
state, we had flows in flow_tree_ but absent in flow_export_info_list_ because
of which flows were not getting aged out.

When flow is enqueued for delete by FlowStats Module to FlowTable module, it
removes entry from flow_export_info_list_ but does not remove from flow_tree_.
The entry from flow_tree_ is removed when the Delete request is sent by FlowMgmt
Module after it is processed by FlowTable module.  If FlowMgmt module sends Add
request instead of Delete (because a FlowEntry which was marked as delete was
added again) we end up in a situation where the two lists of FlowStats module
become out of sync.

Fix:
When Add request is received by FlowStats Module, always add it to
flow_export_info_list_ if it not already present in it. Currently we don’t add
to this list if the entry is already present in flow_tree_.

Added UT

Closes-Bug: #1642830
(cherry picked from commit 01943ed8fa868937086a1f87cd7b009eab326692)

Change-Id: Id6c612b50441d80b277e43fd4ecd748beaad08e7