Do not use configured timer value for rtarget end-of-rib sends and receives
VPN routes are exchanged only after end-of-rib is received for rtarget family
or if the eor-timer for rtarget family expires. When a large eor timer is
configured in order to accomodate large vpn routes, we should not defer the
eor send for rtargets. Otherwise vpn routes exchange will not even get to
start until the large eor timer for route target expires.

Instead always use a sane default of 30 seconds for rtarget eors. This is how
it was before GR code came in

Closes-Bug: #1644033

Change-Id: Id88054929220f8101cee5ed9b34d97045b538220