Add Python code required for Netronome Agilio SmartNIC support
Add Python libraries and unit tests required to make vrouter-port-control
compatible with Netronome Agilio SmartNIC. The unit tests are compatible with
the Python unittest library, and also with testrepository (testr).

Not yet submitted: the main entry point to this code, which is the
vrouter-port-control wrapper script. This is a very simple script that
essentially does the following:

    if __name__ == '__main__':
        import sys

        try:
            from netronome.vrouter.apps import port_control
            sys.exit(port_control.vRouterPortControl().run(argv=sys.argv))
        except KeyboardInterrupt as e:
            print >>sys.stderr, type(e).__name__
            sys.exit(1)

Change-Id: Ief2a3aafe4839e1855be6ac6aac815a58681120d
Partial-Bug: #1644348