Process received xmpp messages always in state-machine task
Before this fix, received messages was processed directly off
TCP read-task callback OnRead(). However while doing so, certain
data structures such as endpoint maps were being examined. This
may not be update because other pending event such as EvOpen may
still be in the queue, yet unprocessed.

Those accesses must be serialized in order to reach correct
conclusion such as finding an already existing XmppConnection
from a given end-point

Fix is to post an EvProcessMessage event to xmpp-state-machine
task and then process the message off the queue asynchronously

Also add a specific test to this this scenario. In the test, we
first disable events processing in the server in XmppStateMachine.
Then we enqueue back-to-back connection requests from the same
agent. Only afterwards, shall we enable the XmppStateMachine events
queue processing and ensure that new connection comes up only after
current one is fully cleaned up.

Verified that test failed (asserted) without the fix and passed
with it.

Change-Id: Ia41f45da8016303e5661307b6243f1dee9b1251a
Closes-Bug: #1643486