* Try ARP request even when EVPN route is in wait for traffic mode
Currently if only EVPN route is in backup mode, GARP packets are
not trapped to agent, hence upon mastership switch agent would
not be able to identify and update EVPN route preference.
Its easily reproducible by changing the group id of VRRP.

Consider VM 1 was configured with group ID 100, but AAP was configured
with AAP vmac of group ID 10, now prefernece of IP route would be
bumped but not EVPN route since mac of GARP packets doesnt match with
AAP configured mac. Now if group ID is changed to 10, VRRP would still
send a GARP but it would not be trapped to agent, since l3 route is
already in high prefernece. Fixing the same by having TRAP flag set even
if EVPN route is in backup mode, also from flow if EVPN route is in low
preference.
Closes-bug:#1644170

Change-Id: I7e9e87f32c453159a56d828bfb16a1dca8627623