Send periodic UVEs in bgp_stress_test also
In control-node, UVEs are send periodically, once a minute. This send code is
not exercised in any of the unit tests. Hence add this to bgp_stress_test to get better test coverage

In bgp_stress_test also, use the same and send UVEs periodically, but a lot
more frequently..

Also, remove triggger to shutdown from introspect. It is not safe to do so
anyways.. One can use service supervisor-control stop to do the same from
command line

Change-Id: Ie646719b0e34ab94bdc766bc0c6cfad9883bcd23
Closes-Bug: #1647514