Agent crash @ peer deletetimeout.
Request enqueued for peer in non deleted state is not processed.
As a solution:
1) config processing has been optimized for handling physical-router
reachability from physical-interface. Due to scaled setup of 4K VRF and 8K
interfaces link walk from physical-interface to physical-device used to be
costly in terms of time taken as for every interface 4K links will be traversed.
To optimize this this link is searched in DB with for the specific
physical-interface associated with VMI and then processed.

2) Route Walker for deletion of path and state for delete peer.
As XMPP channels kept on flapping number of deleted peers created were high.
Each flap creates on deleted peer.
Now every deleted peer will start a walk. So if there are 10 deleted peers, 4K
VRF then number of walks will be 10*4000*5(1 for VRF and 4 for routing tables)
which is i exclusion with DB request processing.
This has been modified to use a single walk which can restart itslef on
finishing off and this walk bunches peers to be processed.

Change-Id: I8e07af8078deccdebc09ba55552a65bd7e766a97
Partial-bug: #1643438