Code for control-node retrieving config via cassandra.
Add code to read one row from cassandra and process the read columns. Also, add
code to convert the incoming column data to a json document that will finally
be fed to the json parser. The conversion is done by the ConfigCass2JsonAdapter
class and handles the quirks in the data stored in cassandra.
Also, add code to read all-rows on startup (bulk sync).

Change-Id: I0ea706554024e314772147fb35c722d008c05fd8
Partial-Bug: #1632470