Added a config know for logical_routers_enabled
When the knob is set to false, schema transformer will not
create 'allow' acl rules for each policy rule. These rules
are required only for the logical router feature to work.
If someone is not using logical routers, turning them off
will reduce the number of ACL rules to half.

Partial-Bug: 1642775

Change-Id: If84f0d7bbb45bc4efd9aaa7b53e8c1fa97cd493f