CNI supported for k8s running inside openstack vm
To support k8s running inside openstack-vm, the interface inside
container is modeled as VMI sub-interface.

CNI creates following interface hierarchy for container,
    <eth0 (inside vm)> <vlan-if (inside-vm)> <macvlan-if (inside container)>

The vlan-interface is used to tag the packets and the macvlan interface
is used to ensure isolation of traffic to/from the containers. The
macvlan interface is created in vepa mode.

The CNI plugin must create vlan-interface and mac-vlan interface
before invoking port-add command to vrouter. However, it needs vlan-tag
before creating vlan-interface. Added a new command to vrouter port-ipc
to get VMI config attributes (vlan and mac-address for the VMI). The VMI
will be addressed based on VM + VN still.

The sequence of operations in CNI is,
    1. Get the VLAN and mac-address for the VMI sub-interface
    2. Create vlan-interface inside openstack-vm
    3. Create macvlan interface inside openstack-vm
    4. Move the macvlan interface inside container
    5. Send port-add command to agent (using port-ipc)
    6. Get the VMI attributes like IP-Address, Netmask, Default-Gw from
       Agent
    7. Program interface inside container with attributes in (6)
    8. Output the results for CNI to STDOUT

Refactored CNI code to use common routines across VEth-Pair and Macvlan
interface types

Change-Id: I6d0e28368b532c16068b510a2f3e715b4ebf5911
Closes-Bug: #1649983